chat and app styling
