+++ src/App.css
@@ -1,5 +1,11 @@
 .App {
   text-align: center;
+  /* background-color: #282c34; */
+  min-height: 100vh;
+}
+
+/* .App {
+  text-align: center;
 }
 
 .App-logo {
@@ -16,7 +22,7 @@
   justify-content: center;
   font-size: calc(10px + 2vmin);
   color: white;
-}
+} */
 
 .joinChatContainer {
   display: flex;
@@ -58,6 +64,10 @@
 .chat-window {
   width: 300px;
   height: 420px;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
 }
 
 .chat-window p {
@@ -182,14 +192,15 @@
   cursor: pointer;
   flex: 15%;
   height: 100%;
-  background: transparent;
+  background-color: #43a047;
   outline: none;
   font-size: 25px;
-  color: lightgray;
+  color: #ffffff;
 }
 
 .chat-window .chat-footer button:hover {
-  color: #43a047;
+  color: #ffffff;
+  background-color: #2e7d32;
 }
 .hide {
   opacity: 0 !important;
